requirements.txt
numpy>=1.26.0
opencv-python>=4.8.0
openvino==2026.0.0
openvino-tokenizers==2026.0.0.0
optimum-intel>=1.15.0
lancedb>=0.6.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0
tqdm>=4.66.0